requirements.txt
python-escpos
requests
PIL
keyboard